videotranscoder/requirements.txt
blinker==1.9.0
click==8.1.8
ffmpeg-python==0.2.0
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
six==1.17.0
sympy==1.13.1
torch==2.6.0
typing_extensions==4.12.2
watchdog==6.0.0
Werkzeug==3.1.3
whisper==1.1.10
